Implemented Section Controller tests.

+++ src/main/java/com/example/antonapi/controller/SectionController.java
@@ -12,6 +12,8 @@ import org.springframework.hateoas.CollectionModel;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.net.URI;
+
 @RepositoryRestController
 @CrossOrigin
 @RequiredArgsConstructor
@@ -29,19 +31,26 @@ public class SectionController {
 
     @GetMapping(path = "/{id}")
     public ResponseEntity<SectionDTO> getSectionById(@PathVariable("id") Long id){
-        return ResponseEntity.ok(sectionModelAssembler.toModel(sectionRepository.getOne(id)));
+        Section foundSection = sectionRepository.findById(id).orElse(null);
+        return foundSection == null
+                ? ResponseEntity.notFound().build()
+                : ResponseEntity.ok(sectionModelAssembler.toModel(foundSection));
     }
 
     @PostMapping
     public ResponseEntity<SectionDTO> addSection(@RequestBody SectionDTO requestSection) {
         Section mappedSection = modelMapper.map(requestSection, Section.class);
         Section addedSection = sectionRepository.saveAndFlush(mappedSection);
-        return ResponseEntity.ok(sectionModelAssembler.toModel(addedSection));
+        SectionDTO returnSectionDTO = sectionModelAssembler.toModel(addedSection);
+        return ResponseEntity.created(URI.create(returnSectionDTO.getLink("self").get().getHref()))
+                .body(returnSectionDTO);
     }
 
     @DeleteMapping(path = "/{id}")
-    public ResponseEntity.HeadersBuilder<?> removeSection(@PathVariable("id") Long id) {
+    public ResponseEntity<?> removeSection(@PathVariable("id") Long id) {
+        if(!sectionRepository.existsById(id))
+            return ResponseEntity.notFound().build();
         sectionRepository.deleteById(id);
-        return ResponseEntity.noContent();
+        return ResponseEntity.noContent().build();
     }
 }
